test: extract apns store coverage
